// src/types/APIsResponseTypes.ts
export type User = {
  id: number;
  firstName: string;
  lastName: string;
  maidenName: string;
  gender: string;
  email: string;
  username: string;
  bloodGroup: string;
  phone: string
  eyeColor: string;
  birthDate: string;
  university: string;
  role: string;
};

export type Product = {
  id: number;
  title: string;
  brand: string;
  category: string;
};